+++ release_info.py
@@ -1,2 +1,2 @@
-VERSION = "5.1"
-TITLE = "Release Review"
+VERSION = "5.3.1"
+TITLE = "Challenge Sharing Fix"